:sparkles: feat: 추천 검색어 캐싱 기능 추가
